Refactor[]: 이동 MIGRATION_1_2
MIGRATION_1_2을 AddContact에서 Contact로 옮김.
